requirements/base.txt
django
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
